requirements.txt
Flask==3.1.0
psycopg2==2.9.10
load_dotenv==0.1.0
dkimpy==1.1.8
msal==1.31.1
requests==2.32.3